requirements.txt
APScheduler==3.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
google-trans-new==1.1.9
grpcio==1.36.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
packaging==20.9
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-telegram-bot==13.3
pytz==2021.1
requests==2.25.1
rfc3986==1.4.0
rsa==4.7.2
six==1.15.0
sniffio==1.2.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.3
